backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
anthropic==0.7.1
python-dotenv==1.0.0
pydantic==2.5.0
cors==1.0.1
mangum==0.17.0
